tests/wpt/web-platform-tests/trusted-types/block-string-assignment-to-HTMLAreaElement-href.tentative.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/helper.sub.js"></script>

  <meta http-equiv="Content-Security-Policy" content="require-trusted-types">
</head>
<body>
  <map>
    <script>
    //helper function for the tests
    function testHref(str, url) {
      var area = document.createElement('area');
      area.href = url;
      area.alt = "Area";
      assert_equals(area.href, str);
    }

    //URL assignments do not throw.
    test(t => {
      testHref(URLS.safe, TrustedURL.create(URLS.safe));
    }, "area.href = URLS.safe, TrustedURL.create");

    test(t => {
      testHref(URLS.safe, TrustedURL.unsafelyCreate(URLS.safe));
    }, "area.href = URLS.safe, TrustedURL.unsafelyCreate");

    // String assignments throw.
    test(t => {
      var area = document.createElement('area');
      assert_throws(new TypeError(), _ => {
        area.href = "A string";
      });
    }, "`area.href = string` throws");

    //Null assignment throws.
    test(t => {
      var area = document.createElement('area');
      assert_throws(new TypeError(), _ => {
        area.href = null;
      });
    }, "`area.href = null` throws");
    </script>
  </map>
